+++ app/src/main/java/woowacourse/paint/mapper/BrushMapper.kt
@@ -3,8 +3,8 @@ package woowacourse.paint.mapper
 import com.now.domain.Brush
 import com.now.domain.BrushColor
 import com.now.domain.BrushWidth
-import woowacourse.paint.custom.uimodel.BrushColorUiModel
-import woowacourse.paint.custom.uimodel.BrushUiModel
+import woowacourse.paint.presentation.uimodel.BrushColorUiModel
+import woowacourse.paint.presentation.uimodel.BrushUiModel
 
 fun Brush.toBrushUiModel(): BrushUiModel {
     return when (this.brushColor) {
@@ -18,10 +18,10 @@ fun Brush.toBrushUiModel(): BrushUiModel {
 
 fun BrushUiModel.toBrush(): Brush {
     return when (this.brushColor) {
-        BrushColorUiModel.RED -> { Brush(BrushColor.RED, BrushWidth(this.brushWidth)) }
-        BrushColorUiModel.YELLOW -> { Brush(BrushColor.YELLOW, BrushWidth(this.brushWidth)) }
-        BrushColorUiModel.ORANGE -> { Brush(BrushColor.ORANGE, BrushWidth(this.brushWidth)) }
-        BrushColorUiModel.GREEN -> { Brush(BrushColor.GREEN, BrushWidth(this.brushWidth)) }
-        BrushColorUiModel.BLUE -> { Brush(BrushColor.BLUE, BrushWidth(this.brushWidth)) }
+        woowacourse.paint.presentation.uimodel.BrushColorUiModel.RED -> { Brush(BrushColor.RED, BrushWidth(this.brushWidth)) }
+        woowacourse.paint.presentation.uimodel.BrushColorUiModel.YELLOW -> { Brush(BrushColor.YELLOW, BrushWidth(this.brushWidth)) }
+        woowacourse.paint.presentation.uimodel.BrushColorUiModel.ORANGE -> { Brush(BrushColor.ORANGE, BrushWidth(this.brushWidth)) }
+        woowacourse.paint.presentation.uimodel.BrushColorUiModel.GREEN -> { Brush(BrushColor.GREEN, BrushWidth(this.brushWidth)) }
+        woowacourse.paint.presentation.uimodel.BrushColorUiModel.BLUE -> { Brush(BrushColor.BLUE, BrushWidth(this.brushWidth)) }
     }
 }
