+++ app/src/main/java/woowacourse/paint/presentation/uimodel/BrushColorUiModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint.custom.uimodel
+package woowacourse.paint.presentation.uimodel
 
 import androidx.annotation.ColorInt
