+++ app/src/main/java/woowacourse/paint/presentation/uimodel/BrushUiModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint.custom.uimodel
+package woowacourse.paint.presentation.uimodel
 
 data class BrushUiModel(
     val brushColor: BrushColorUiModel,
